+++ tests/LookupEngine.Tests.Benchmarks/Benchmarks/FormatMemberNameBenchmark.cs
@@ -15,7 +15,7 @@
 using System.Reflection;
 using BenchmarkDotNet.Attributes;
 
-namespace LookupEngine.Tests.Performance.Benchmarks;
+namespace LookupEngine.Tests.Benchmarks.Benchmarks;
 
 /// <summary>
 ///     Compares strategies for the <c>Name (Type1, ref Type2, ...)</c> member name format, as implemented in <c>ReflexionFormater.FormatMemberName</c>.